requirements.txt
matplotlib==3.1.3
pytest==6.1.2
numpy==1.18.1
PyYAML==5.3.1
